Add Movies data

+++ app/src/main/java/in/khofid/moviecatalogue/data/Movie.kt
@@ -14,5 +14,5 @@ data class Movie(
     @SerializedName("backdrop_path")
     val backdrop: String,
     @SerializedName("vote_average")
-    val vote: Double
+    val vote: Float
 )
